add task to clear puppeteer windows
The `before` action which sets up metamask is executed by cypress for
each new file, however the window state remains between runs. This
caused tests to hang on the unlock screen. By clearing the windows in
an `after` block the tests will appropriate unlock metamask and resume
the cypress test.

Add a `clearWindows()` task that sets both windows to null.